restructure some files
