--- eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceOperationKind.cs
@@ -1,15 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Azure.Generator.Management.Models
-{
-    internal enum ResourceOperationKind
-    {
-        Action,
-        Create,
-        Delete,
-        Read,
-        List,
-        Update,
-    }
-}
