+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceSchemas/ResourceOperationKindExtensions.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Generator.Management.Models;
+
+internal static class ResourceOperationKindExtensions
+{
+    public static bool IsCrudKind(this ResourceOperationKind kind)
+    {
+        return kind == ResourceOperationKind.Create ||
+               kind == ResourceOperationKind.Read ||
+               kind == ResourceOperationKind.Update ||
+               kind == ResourceOperationKind.Delete;
+    }
+}
