+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceSchemas/ResourceScope.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Generator.Management.Models;
+
+internal enum ResourceScope
+{
+    Tenant,
+    Subscription,
+    ResourceGroup,
+    ManagementGroup,
+    Extension,
+}
